Refactor tools into separate modules.
