+++ kafka-daq-report/src/message_debug.rs
@@ -1,0 +1,65 @@
+use super::CommonOpts;
+use anyhow::Result;
+use rdkafka::{
+    consumer::{CommitMode, Consumer, StreamConsumer},
+    Message,
+};
+use supermusr_streaming_types::dat1_digitizer_analog_trace_v1_generated::{
+    digitizer_analog_trace_message_buffer_has_identifier, root_as_digitizer_analog_trace_message,
+};
+use tracing::{debug, info, warn};
+
+// Message dumping tool
+pub(crate) async fn run(args: CommonOpts) -> Result<()> {
+    tracing_subscriber::fmt::init();
+
+    let consumer: StreamConsumer = supermusr_common::generate_kafka_client_config(
+        &args.broker,
+        &args.username,
+        &args.password,
+    )
+    .set("group.id", &args.consumer_group)
+    .set("enable.partition.eof", "false")
+    .set("session.timeout.ms", "6000")
+    .set("enable.auto.commit", "false")
+    .create()?;
+
+    consumer.subscribe(&[&args.trace_topic])?;
+
+    loop {
+        match consumer.recv().await {
+            Err(e) => warn!("Kafka error: {}", e),
+            Ok(msg) => {
+                debug!(
+                    "key: '{:?}', topic: {}, partition: {}, offset: {}, timestamp: {:?}",
+                    msg.key(),
+                    msg.topic(),
+                    msg.partition(),
+                    msg.offset(),
+                    msg.timestamp()
+                );
+
+                if let Some(payload) = msg.payload() {
+                    if digitizer_analog_trace_message_buffer_has_identifier(payload) {
+                        match root_as_digitizer_analog_trace_message(payload) {
+                            Ok(data) => {
+                                info!(
+                                    "Trace packet: dig. ID: {}, metadata: {:?}",
+                                    data.digitizer_id(),
+                                    data.metadata()
+                                );
+                            }
+                            Err(e) => {
+                                warn!("Failed to parse message: {}", e);
+                            }
+                        }
+                    } else {
+                        warn!("Unexpected message type on topic \"{}\"", msg.topic());
+                    }
+                }
+
+                consumer.commit_message(&msg, CommitMode::Async).unwrap();
+            }
+        };
+    }
+}
